\name{boa-internal}
\alias{boa.getinput}
\alias{boa.menu.analysis}
\alias{boa.menu.chains}
\alias{boa.menu.coda}
\alias{boa.menu.data}
\alias{boa.menu.file}
\alias{boa.menu.import}
\alias{boa.menu.par}
\alias{boa.menu.parms}
\alias{boa.menu.plot}
\alias{boa.menu.plotcoda}
\alias{boa.menu.plotdesc}
\alias{boa.menu.setpar}
\alias{boa.menu.stats}
\alias{boa.menu.window}

\title{Internal BOA Functions}
\description{Internal BOA functions}

\usage{
boa.getinput(message, n = 1, evaluate = TRUE)
boa.menu.analysis()
boa.menu.chains()
boa.menu.coda()
boa.menu.data()
boa.menu.file()
boa.menu.import()
boa.menu.par()
boa.menu.parms()
boa.menu.plot()
boa.menu.plotcoda()
boa.menu.plotdesc()
boa.menu.setpar(group)
boa.menu.stats()
boa.menu.window()
}

\details{These are not to be called by the user}
\keyword{internal}
