\name{boa.init}
\alias{boa.init}

\title{Start Session}
\description{
  boa.init() is the first function called to start a BOA session when using the
  command-line. It sets up the internal data structures and initializes them
  with the appropriate default values.
}
\usage{
boa.init()
}

\author{Brian J. Smith}

\seealso{ \code{\link{boa.load}}, \code{\link{boa.quit}} }

\keyword{utilities}
