\name{boa.plot.par}
\alias{boa.plot.par}

\title{Set Plotting Parameters}
\description{
  Set Plotting Parameters
}
\usage{
boa.plot.par(mfdim = c(1, 1))
}

\arguments{
  \item{mfdim}{Numeric vector with two elements giving the number of rows and 
  column, respectively, of plots to display on the page. Plots will be drawn 
  row-by-row.}
}

\author{Brian J. Smith}

\keyword{utilities}
