\name{boa.plot.title}
\alias{boa.plot.title}

\title{Add Plot Title}
\description{
  Adds a title to the graphics window. This function should be called after all 
  plots have been added to the target graphics window.
}
\usage{
boa.plot.title(text)
}

\arguments{
  \item{text}{Character string giving the title to be centered and displayed 
  across the top of the graphics window.}
}

\author{Brian J. Smith}

\keyword{utilities}
