\name{boa.print.cor}
\alias{boa.print.cor}

\title{Print Correlation Matrices}
\description{
  Iteratively computes and displays the correlation matrices for the parameters 
  in each of the MCMC sequences in the working session list of sequences.
}
\usage{
boa.print.cor()
}

\author{Brian J. Smith}

\keyword{print}
