\name{boa.version}
\alias{boa.version}

\title{Version Information}
\description{
  Information about this release of BOA.
}
\usage{
boa.version()
}

\value{
  \item{name}{Character string giving the program name.}
  \item{major}{Major number of this release.}
  \item{minor}{Minor number of this release.}
  \item{revision}{Revision number for this version.}
  \item{system}{Character string giving the platform for which this version 
  of BOA was designed.}
}

\author{Brian J. Smith}

\seealso{ \code{\link{boa.license}} }

\keyword{internal}
