\name{colVars}
\alias{colVars}

\title{Column Variances}
\description{
  Computes variances by column.
}
\usage{
colVars(x, na.rm = FALSE, unbiased = TRUE, SumSquares = FALSE)
}

\arguments{
  \item{x}{A matrix, array, or data frame}
  \item{na.rm}{Logical value indicating that missing values be omitted from the
  calculations.}
  \item{unbiased}{Not implemented}
  \item{SumSquares}{Not implemented}
}

\keyword{internal}
